requirements.txt
absl-py==2.0.0
antlr4-python3-runtime==4.9.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.6
dlib==19.22.1
editdistance==0.6.0
fairseq @ file:///home/user/workspace_tcx/av_hubert/fairseq
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
google-auth==2.24.0
google-auth-oauthlib==1.0.0
grpcio==1.59.3
hydra-core==1.3.2
idna==3.6
imageio==2.33.0
importlib-metadata==6.9.0
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.5.4.60
packaging==23.2
Pillow==10.1.0
portalocker==2.8.2
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-speech-features==0.6
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.3.2
scikit-image==0.17.2
scikit-video==1.1.11
scipy==1.5.4
sentencepiece==0.1.96
six==1.16.0
submitit==1.4.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
torch==2.1.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
